+++ R/bullets.R
@@ -16,6 +16,7 @@
 #' * No name: Item without a prefix.
 #' * ` `: Indented item.
 #' * `*`: Item with a bullet.
+#' * `>`: Item with an arrow or pointer.
 #' * `v`: Item with a green "tick" symbol, like [cli_alert_success()].
 #' * `x`: Item with a ref cross, like [cli_alert_danger()].
 #' * `!`: Item with a yellow exclamation mark, like [cli_alert_warning()].
@@ -32,17 +33,18 @@
 #'
 #' @export
 #' @examples
-#' cli_memo(c(
+#' cli_bullets(c(
 #'         "noindent",
 #'   " " = "indent",
 #'   "*" = "bullet",
+#'   ">" = "arrow",
 #'   "v" = "success",
 #'   "x" = "danger",
 #'   "!" = "warning",
 #'   "i" = "info"
 #' ))
 
-cli_memo <- function(text, id = NULL, class = NULL,
+cli_bullets <- function(text, id = NULL, class = NULL,
                      .envir = parent.frame()) {
   cli__message(
     "memo",