+++ R/cliapp.R
@@ -11,20 +11,8 @@ cliapp <- function(theme = getOption("cli.theme"),
 
     ## Meta
     meta = function(...) {
-      old <- app$output
-      on.exit(app$output <- old, add = TRUE)
-      on.exit(app$signal <- NULL, add = TRUE)
-      out <- rawConnection(raw(1000), open = "w")
-      on.exit(close(out), add = TRUE)
-      app$output <- out
-      app$signal <- FALSE
-
-      for (msg in list(...)) {
-        do.call(app[[msg$type]], msg$args)
-      }
-
-      txt <- rawToChar(rawConnectionValue(out))
-      clii__message(txt, appendLF = FALSE, output = old, signal = TRUE)
+      txt <- cli__fmt(list(...), collapse = TRUE, app = app)
+      clii__message(txt, appendLF = FALSE, output = app$output, signal = TRUE)
     },
 
     ## Themes
@@ -124,7 +112,7 @@ cliapp <- function(theme = getOption("cli.theme"),
     output = NULL,
 
     get_current_style = function()
-      tail(app$styles, 1)[[1]],
+      utils::tail(app$styles, 1)[[1]],
 
     xtext = function(text = NULL, .list = NULL, indent = 0, padding = 0)
       clii__xtext(app, text, .list = .list, indent = indent,
@@ -254,19 +242,20 @@ clii_alert <- function(app, type, text, id, class, wrap) {
   clii__container_start(app, "div", id = id,
                        class = paste(class, "alert", type))
   on.exit(clii__container_end(app, id), add = TRUE)
-  text <- app$inline(text)
-  style <- app$get_current_style()
-  before <- call_if_fun(style$before) %||% ""
-  after <- call_if_fun(style$after) %||% ""
-  before <- gsub(" ", "\u00a0", before)
-  after <- gsub(" ", "\u00a0", after)
-  text[1] <- paste0(before, text[1])
-  text[length(text)] <- paste0(text[length(text)], after)
-  if (is.function(style$fmt)) text <- style$fmt(text)
   if (wrap) {
-    text <- ansi_strwrap(text, exdent = ansi_nchar(before, "width"))
+    app$xtext(text)
+  } else {
+    text <- app$inline(text)
+    style <- app$get_current_style()
+    before <- call_if_fun(style$before) %||% ""
+    after <- call_if_fun(style$after) %||% ""
+    before <- gsub(" ", "\u00a0", before)
+    after <- gsub(" ", "\u00a0", after)
+    text[1] <- paste0(before, text[1])
+    text[length(text)] <- paste0(text[length(text)], after)
+    if (is.function(style$fmt)) text <- style$fmt(text)
+    app$cat_ln(text)
   }
-  app$cat_ln(text)
 }
 
 ## Memo -------------------------------------------------------------
@@ -281,7 +270,7 @@ clii_memo <- function(app, text, id, class) {
   nms[is.na(nms) | nms == ""] <- "empty"
   nms[nms == " "] <- "space"
   nms <- gsub(" ", "-", nms)
-  cls <- paste0("memo-item-", nms)
+  cls <- paste0("memo-item memo-item-", nms)
 
   lapply(seq_along(text), function(i) {
     iid <- new_uuid()