+++ R/tree.R
@@ -114,7 +114,7 @@ tree <- function(data, root = data[[1]][[1]], style = NULL,
   style <- style %||% box_chars()
 
   labels <- if (ncol(data) >= 3) data[[3]] else data[[1]]
-  trimlabs <- data$trimmed %||% labels
+  trimlabs <- data[["trimmed"]] %||% labels
 
   seen <- character()
   res <- character()
@@ -183,9 +183,7 @@ box_chars <- function() {
   }
 }
 
-#' @importFrom methods setOldClass
-
-setOldClass(c("tree", "character"))
+methods::setOldClass(c("tree", "character"))
 
 #' @export