+++ NEWS.md
@@ -1,4 +1,28 @@
 
+# cli 2.5.0
+
+* New `style_no_*()` functions to locally undo styling.
+  New `col_none()` and `bg_none()` functions to locally undo text color
+  and background color.
+
+* It is now possible to undo text and background color in a theme, by
+  setting them to `NULL` or `"none"`.
+
+* `cli_memo()` was renamed to `cli_bullets()`, as it is by default
+  formatted as a bullet list (#250).
+
+* New `ansi_toupper()`, `ansi_tolower` and `ansi_chartr()` functions,
+  the ANSI styling aware variants of `toupper()`, `tolower()` and
+  `chartr()` (#248).
+
+* New `test_that_cli()` helper function to write testthat tests for
+  cli output.
+
+* `tree()` now does not produce warnings for tibbles (#238).
+
+* New inline style: `.cls` to format class names, e.g.
+  `"{.var fit} must be an {.cls lm} object"`.
+
 # cli 2.4.0
 
 * New `cli_memo()` function to create a list of items or tasks.